1. Tentatively fixes #6. This is with the completion of multidog(). 2. Added a new vignette about multidog(). 3. Deleted the old vignette about parallel computing, as folks should now use multidog(). 4. Coerce signed objects to be unsigned prior to comparison. This should get rid of all of those compiler warnings I get whenever I install updog.

+++ src/em.cpp
@@ -36,7 +36,7 @@ double wem_obj(const arma::vec & pivec,
                const long double & lambda) {
   arma::vec lpi = lmat.t() * pivec;
   double obj = 0.0;
-  for (int k = 0; k < weight_vec.n_elem; k++) {
+  for (int k = 0; (unsigned)k < weight_vec.n_elem; k++) {
     if ((weight_vec(k) > TOL) && (lpi(k) > TOL)) {
       obj = obj + weight_vec(k) * std::log(lpi(k));
     } else if ((weight_vec(k) > TOL * 1000) && (lpi(k) < TOL)) {
